I05-70 Fix primary axis tagging when a locationMap is used.
Previously the primary axis tagging was handled by the NDW
makeScannablesAndMonitors method however if all the scannables have
already been handled by the locationMap no primary tag will be added and
if a subset are handled the primary tag may be on the wrong axis.

Due to the abstraction of the writing this requires a flag to be passed
through quite  a few classes.